+++ bld/sdk/wresedit/waccel/c/wmain.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2022 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2023 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -417,7 +417,7 @@ char *WCreateEditTitle( WAccelEditInfo *einfo )
     char        *title;
     char        *fname;
     char        *text;
-    int         offset;
+    size_t      offset;
     size_t      len;
 
     title = NULL;
